requirements.txt
# Core ML libraries
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.2

# Data validation & generation
faker==30.0.0
pyyaml==6.0.2
sqlalchemy==2.0.34

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# Model persistence & APIs
flask==3.0.3
joblib==1.4.2
python-dotenv==1.0.1

# Optional explainability tools
shap==0.46.0

# Utility tools
tqdm==4.66.5
